Move scalar spec in config
